+++ build-plugin/_deps/juce-build/modules/cmake_install.cmake
@@ -136,6 +136,6 @@ endif()
 string(REPLACE ";" "\n" CMAKE_INSTALL_MANIFEST_CONTENT
        "${CMAKE_INSTALL_MANIFEST_FILES}")
 if(CMAKE_INSTALL_LOCAL_ONLY)
-  file(WRITE "/Users/michael/apps/sappkeys/build-plugin/_deps/juce-build/modules/install_local_manifest.txt"
+  file(WRITE "/Users/michael/apps/sappkeys/.claude/worktrees/admiring-noether-77a6b7/build-plugin/_deps/juce-build/modules/install_local_manifest.txt"
      "${CMAKE_INSTALL_MANIFEST_CONTENT}")
 endif()